requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=1.5.0
tqdm>=4.64.0
python-dotenv>=1.0.0